update users api stuff
